# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Daily Expense Tracker API
 * OpenAPI spec version: 0.1.0
 */
export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface ErrorEnvelope {
  error: string;
}

export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Expense {
  id: number;
  /** Amount in dollars */
  amount: number;
  /** Expense category (e.g. Food, Transport, Shopping) */
  category: string;
  description: string;
  /** ISO date string (YYYY-MM-DD) */
  date: string;
  /**
     * Set when this expense was auto-logged from a recurring rule
     * @nullable
     */
  recurringId: number | null;
  /** ISO datetime string */
  createdAt: string;
}

export interface ExpenseInput {
  /** @minimum 0.01 */
  amount: number;
  /** @minLength 1 */
  category: string;
  /** @minLength 1 */
  description: string;
  /** ISO date string (YYYY-MM-DD) */
  date: string;
}

export interface ExpenseUpdate {
  /** @minimum 0.01 */
  amount?: number;
  /** @minLength 1 */
  category?: string;
  /** @minLength 1 */
  description?: string;
  date?: string;
}

export interface Budget {
  id: number;
  /** Daily spending limit in dollars */
  dailyLimit: number;
  /** Monthly spending limit in dollars */
  monthlyLimit: number;
  /** ISO datetime string */
  updatedAt: string;
}

export interface BudgetInput {
  /** @minimum 1 */
  dailyLimit: number;
  /** @minimum 1 */
  monthlyLimit: number;
}

export interface CategoryBreakdown {
  category: string;
  total: number;
  count: number;
  percentage: number;
}

export interface DailySummary {
  date: string;
  totalSpent: number;
  dailyLimit: number;
  remaining: number;
  percentUsed: number;
  expenseCount: number;
  categories: CategoryBreakdown[];
}

export interface DaySpending {
  date: string;
  total: number;
  count: number;
}

export interface WeeklySummary {
  weekTotal: number;
  dailyAverage: number;
  /** @nullable */
  topCategory: string | null;
  days: DaySpending[];
}

export interface SpendingTip {
  id: string;
  /** alert | tip | warning | positive */
  type: string;
  title: string;
  message: string;
}

export interface SpendingInsights {
  tips: SpendingTip[];
  alerts: SpendingTip[];
}

export type RecurringExpenseFrequency = typeof RecurringExpenseFrequency[keyof typeof RecurringExpenseFrequency];


export const RecurringExpenseFrequency = {
  daily: 'daily',
  weekly: 'weekly',
  monthly: 'monthly',
} as const;

export interface RecurringExpense {
  id: number;
  description: string;
  amount: number;
  category: string;
  frequency: RecurringExpenseFrequency;
  /** ISO date string (YYYY-MM-DD); anchors the repeat schedule */
  startDate: string;
  active: boolean;
  /**
     * Occurrences up to this date have been auto-logged
     * @nullable
     */
  lastMaterializedDate: string | null;
  /** ISO datetime string */
  createdAt: string;
}

export type RecurringExpenseInputFrequency = typeof RecurringExpenseInputFrequency[keyof typeof RecurringExpenseInputFrequency];


export const RecurringExpenseInputFrequency = {
  daily: 'daily',
  weekly: 'weekly',
  monthly: 'monthly',
} as const;

export interface RecurringExpenseInput {
  /** @minLength 1 */
  description: string;
  /** @minimum 0.01 */
  amount: number;
  /** @minLength 1 */
  category: string;
  frequency: RecurringExpenseInputFrequency;
  /** ISO date string (YYYY-MM-DD) */
  startDate: string;
  active?: boolean;
}

export type RecurringExpenseUpdateFrequency = typeof RecurringExpenseUpdateFrequency[keyof typeof RecurringExpenseUpdateFrequency];


export const RecurringExpenseUpdateFrequency = {
  daily: 'daily',
  weekly: 'weekly',
  monthly: 'monthly',
} as const;

export interface RecurringExpenseUpdate {
  /** @minLength 1 */
  description?: string;
  /** @minimum 0.01 */
  amount?: number;
  /** @minLength 1 */
  category?: string;
  frequency?: RecurringExpenseUpdateFrequency;
  startDate?: string;
  active?: boolean;
}

export type PreferencesCurrency = typeof PreferencesCurrency[keyof typeof PreferencesCurrency];


export const PreferencesCurrency = {
  USD: 'USD',
  EUR: 'EUR',
  GBP: 'GBP',
  JPY: 'JPY',
  INR: 'INR',
  CAD: 'CAD',
  AUD: 'AUD',
} as const;

export interface Preferences {
  id: number;
  currency: PreferencesCurrency;
  /** Percent of daily limit at which warnings fire */
  alertThreshold: number;
  /** ISO datetime string */
  updatedAt: string;
}

export type PreferencesInputCurrency = typeof PreferencesInputCurrency[keyof typeof PreferencesInputCurrency];


export const PreferencesInputCurrency = {
  USD: 'USD',
  EUR: 'EUR',
  GBP: 'GBP',
  JPY: 'JPY',
  INR: 'INR',
  CAD: 'CAD',
  AUD: 'AUD',
} as const;

export interface PreferencesInput {
  currency: PreferencesInputCurrency;
  /**
     * @minimum 50
     * @maximum 100
     */
  alertThreshold: number;
}

export interface SpendingStats {
  date: string;
  monthToDate: number;
  /** Straight-line projection of this month's spend */
  projectedMonthEnd: number;
  monthlyLimit: number;
  monthPercentUsed: number;
  avgPerDay: number;
  daysElapsed: number;
  daysInMonth: number;
  /** Consecutive days (ending today) at or under the daily limit */
  underBudgetStreak: number;
  totalExpenseCount: number;
  activeRecurringCount: number;
  /** Approximate monthly cost of all active recurring rules */
  recurringMonthlyTotal: number;
}

/**
 * Opaque session token — `Bearer <sid>`.
 */
export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
/**
 * Relative path to redirect to after login (must start with `/`). Defaults to `/`.
 */
returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
code?: string;
state?: string;
iss?: string;
};

export type LogoutBrowserSessionParams = {
returnTo?: string;
};

export type ListExpensesParams = {
/**
 * ISO date string (YYYY-MM-DD)
 */
startDate?: string;
/**
 * ISO date string (YYYY-MM-DD)
 */
endDate?: string;
category?: string;
};

export type GetDailySummaryParams = {
/**
 * ISO date string (YYYY-MM-DD), defaults to today
 */
date?: string;
};

export type GetWeeklySummaryParams = {
/**
 * ISO date string (YYYY-MM-DD); anchors the week, defaults to today
 */
date?: string;
};

export type GetSpendingStatsParams = {
/**
 * ISO date string (YYYY-MM-DD), defaults to today
 */
date?: string;
};

export type GetSpendingTipsParams = {
/**
 * ISO date string (YYYY-MM-DD) treated as "today", defaults to server date
 */
date?: string;
};